+++ LowLevelConnector/BlindConnector.h
@@ -3,14 +3,14 @@
 #include <Arduino.h>
 class BlindConnector {
 public:
-  BlindConnector(int writePin, int blindPin, int luxPin);
-  void writePWM(int val);
-  int readBlindAnalog();
-  int readLuxAnalog();
+  BlindConnector(uint8_t writePin, uint8_t blindPin, uint8_t luxPin);
+  void writePWM(uint8_t val);
+  uint16_t readBlindAnalog();
+  uint16_t readLuxAnalog();
 private:
-  int _writePin;
-  int _luxPin;
-  int _blindPin;
-  int _clipPWM(int val);
+  uint8_t _writePin;
+  uint8_t _luxPin;
+  uint8_t _blindPin;
+  uint8_t _clipPWM(uint8_t val);
 };
 #endif
